+++ src/op3/__init__.py
@@ -1,3 +1,3 @@
 """op3 package."""
 
-__version__ = "0.1.8"
+__version__ = "0.1.9"