+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.9] - 2026-04-21
+
+### Docs
+- Update README.md
+
+### Test
+- Update tests/unit/test_build_scanner.py
+- Update tests/unit/test_probe_registry.py
+
 ## [0.1.8] - 2026-04-21
 
 ### Docs